refactor stringcache
