+++ polars/polars-core/src/chunked_array/logical/categorical/mod.rs
@@ -2,6 +2,7 @@ mod builder;
 mod from;
 mod merge;
 mod ops;
+pub mod stringcache;
 
 use super::*;
 use crate::prelude::*;
