+++ polars/polars-core/src/lib.rs
@@ -24,20 +24,14 @@ mod tests;
 pub(crate) mod vector_hasher;
 use once_cell::sync::Lazy;
 
-#[cfg(any(feature = "dtype-categorical", feature = "object"))]
+#[cfg(feature = "object")]
 use std::time::{SystemTime, UNIX_EPOCH};
 
-#[cfg(feature = "dtype-categorical")]
-use ahash::AHashMap;
 use rayon::{ThreadPool, ThreadPoolBuilder};
-#[cfg(feature = "dtype-categorical")]
-use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Mutex;
-#[cfg(feature = "dtype-categorical")]
-use std::sync::MutexGuard;
 
 #[cfg(feature = "dtype-categorical")]
-use smartstring::{LazyCompact, SmartString};
+pub use crate::chunked_array::logical::categorical::stringcache::*;
 
 #[cfg(feature = "object")]
 pub(crate) static PROCESS_ID: Lazy<u128> = Lazy::new(|| {
@@ -63,89 +57,5 @@ pub static POOL: Lazy<ThreadPool> = Lazy::new(|| {
         .expect("could not spawn threads")
 });
 
-#[cfg(feature = "dtype-categorical")]
-struct SCacheInner {
-    map: AHashMap<SmartString<LazyCompact>, u32>,
-    uuid: u128,
-}
-
-#[cfg(feature = "dtype-categorical")]
-impl Default for SCacheInner {
-    fn default() -> Self {
-        Self {
-            map: Default::default(),
-            uuid: SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap()
-                .as_nanos(),
-        }
-    }
-}
-
-/// Used by categorical data that need to share global categories.
-/// In *eager* you need to specifically toggle global string cache to have a global effect.
-/// In *lazy* it is toggled on at the start of a computation run and turned of (deleted) when a
-/// result is produced.
-#[cfg(feature = "dtype-categorical")]
-pub(crate) struct StringCache(pub(crate) Mutex<SCacheInner>);
-
-#[cfg(feature = "dtype-categorical")]
-impl StringCache {
-    pub(crate) fn lock_map(&self) -> MutexGuard<SCacheInner> {
-        self.0.lock().unwrap()
-    }
-
-    pub(crate) fn clear(&self) {
-        let mut lock = self.lock_map();
-        *lock = Default::default();
-    }
-}
-
-#[cfg(feature = "dtype-categorical")]
-impl Default for StringCache {
-    fn default() -> Self {
-        StringCache(Mutex::new(Default::default()))
-    }
-}
-
-#[cfg(feature = "dtype-categorical")]
-pub(crate) static USE_STRING_CACHE: AtomicBool = AtomicBool::new(false);
-
-#[cfg(feature = "dtype-categorical")]
-pub(crate) static STRING_CACHE: Lazy<StringCache> = Lazy::new(Default::default);
-
 // utility for the tests to ensure a single thread can execute
 pub static SINGLE_LOCK: Lazy<Mutex<()>> = Lazy::new(|| Mutex::new(()));
-
-#[cfg(feature = "dtype-categorical")]
-pub fn with_string_cache<F: FnOnce() -> T, T>(func: F) -> T {
-    toggle_string_cache(true);
-    let out = func();
-    toggle_string_cache(false);
-    out
-}
-
-/// Use a global string cache for the Categorical Types.
-///
-/// This is used to cache the string categories locally.
-/// This allows join operations on categorical types.
-#[cfg(feature = "dtype-categorical")]
-pub fn toggle_string_cache(toggle: bool) {
-    USE_STRING_CACHE.store(toggle, Ordering::Release);
-
-    if !toggle {
-        STRING_CACHE.clear()
-    }
-}
-
-/// Reset the global string cache used for the Categorical Types.
-#[cfg(feature = "dtype-categorical")]
-pub fn reset_string_cache() {
-    STRING_CACHE.clear()
-}
-
-/// Check if string cache is set.
-#[cfg(feature = "dtype-categorical")]
-pub(crate) fn use_string_cache() -> bool {
-    USE_STRING_CACHE.load(Ordering::Acquire)
-}
